+++ packages/orq-rc/RELEASES.md
@@ -6359,3 +6359,13 @@ Based on:
 - [typescript v4.3.0-rc.20] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.20 - packages/orq-rc
+
+## 2026-02-03 13:57:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.21 - packages/orq-rc